+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/domain/SurveillanceSearchOptions.java
@@ -1,0 +1,5 @@
+package gov.healthit.chpl.domain;
+
+public enum SurveillanceSearchOptions {
+	OPEN_SURVEILLANCE, CLOSED_SURVEILLANCE, OPEN_NONCONFORMITY, CLOSED_NONCONFORMITY;
+}